[#3913] Move summoning functionality into `SummonActivity`
Shifts the summoning behavior from `SummonData` into
`SummonActivity` and adds a modified dialog to select summoning
profile, sizes, and types.

Adds `button` options to `ActivityUseDialog` for customizing the
buttons that appear and `display` to control what sections are
displayed or not.

Closes #3913
