+++ lang/en.json
@@ -2283,6 +2283,7 @@
 "DND5E.SubclassMismatchWarn": "{name} subclass has no matching class with identifier '{class}'.",
 "DND5E.SubclassName": "Subclass Name",
 "DND5E.Subtype": "Subtype",
+
 "DND5E.SUMMON": {
   "Title": "Summon",
   "FIELDS": {
@@ -2390,6 +2391,11 @@
     "Profiles": "Profiles",
     "Summoning": "Summoning"
   },
+  "Action": {
+    "Place": "Place Summons",
+    "Summon": "Summon",
+    "View": "View Summon"
+  },
   "CreatureChanges": {
     "Label": "Creature Changes",
     "Hint": "Changes that will be made to the creature being summoned. Any @ references used in the following formulas will be based on the summoner's stats. Summoned creature's stats can be referenced using @summon (e.g. @summon.attributes.hd.max to reference the creature's hit dice count)."
@@ -2399,25 +2405,15 @@
     "Hint": "Changes made to items on the summoned creature."
   },
   "Profile": {
+    "Label": "Summons Profile",
     "Action": {
       "Create": "Create Profile",
       "Delete": "Delete Profile"
     },
+    "ChallengeRatingLabel": "Challenge Rating of {cr} or lower",
     "DropHint": "Drop creature here",
     "Empty": "Click below to create a profile.",
     "EmptyDrop": "Click below to create a profile or drop a creature to summon here."
-  }
-},
-"DND5E.Summoning": {
-  "Label": "Summoning",
-  "Action": {
-    "Place": "Place Summons",
-    "Summon": "Summon",
-    "View": "View Summon"
-  },
-  "Profile": {
-    "Label": "Summons Profile",
-    "ChallengeRatingLabel": "Challenge Rating of {cr} or lower"
   },
   "Warning": {
     "CreateActor": "You must have 'Create New Actors' permission in order to summon directly from a compendium.",
@@ -2427,6 +2423,10 @@
     "NoProfile": "Cannot find summoning profile {profileId} on '{item}'.",
     "Wildcard": "You must have 'Use File Browser' permissions to summon creatures with wildcard artwork."
   }
+},
+
+"DND5E.Summoning": {
+  "Label": "Summoning"
 },
 "DND5E.Supply": "Supply",
 "DND5E.Suppressed": "Suppressed",
