+++ module/applications/activity/_module.mjs
@@ -8,3 +8,4 @@ export {default as UtilitySheet} from "./utility-sheet.mjs";
 
 export {default as ActivityUsageDialog} from "./activity-usage-dialog.mjs";
 export {default as EnchantUsageDialog} from "./enchant-usage-dialog.mjs";
+export {default as SummonUsageDialog} from "./summon-usage-dialog.mjs";
