+++ module/applications/item/ability-use-dialog.mjs
@@ -65,7 +65,6 @@ export default class AbilityUseDialog extends Dialog {
       item,
       ...config,
       slotOptions: config.consumeSpellSlot ? slotOptions : [],
-      summoningOptions: this._createSummoningOptions(item),
       resourceOptions: resourceOptions,
       resourceArray: Array.isArray(resourceOptions),
       concentration: {
@@ -179,47 +178,6 @@ export default class AbilityUseDialog extends Dialog {
       }
     }
 
-    return options;
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Create details on the summoning profiles and other related options.
-   * @param {Item5e} item  The item.
-   * @returns {{ profiles: object, creatureTypes: object }|null}
-   */
-  static _createSummoningOptions(item) {
-    const summons = item.system.summons;
-    if ( !summons?.profiles.length ) return null;
-    const options = { mode: summons.mode };
-    const rollData = item.getRollData();
-    const level = summons.relevantLevel;
-    options.profiles = Object.fromEntries(
-      summons.profiles
-        .map(profile => {
-          if ( !summons.mode && !fromUuidSync(profile.uuid) ) return null;
-          const withinRange = ((profile.level.min ?? -Infinity) <= level) && (level <= (profile.level.max ?? Infinity));
-          if ( !withinRange ) return null;
-          return [profile._id, summons.getProfileLabel(profile, rollData)];
-        })
-        .filter(f => f)
-    );
-    if ( Object.values(options.profiles).every(p => p.startsWith("1 × ")) ) {
-      Object.entries(options.profiles).forEach(([k, v]) => options.profiles[k] = v.replace("1 × ", ""));
-    }
-    if ( Object.values(options.profiles).length <= 1 ) {
-      options.profile = Object.keys(options.profiles)[0];
-      options.profiles = null;
-    }
-    if ( summons.creatureSizes.size > 1 ) options.creatureSizes = summons.creatureSizes.reduce((obj, k) => {
-      obj[k] = CONFIG.DND5E.actorSizes[k]?.label;
-      return obj;
-    }, {});
-    if ( summons.creatureTypes.size > 1 ) options.creatureTypes = summons.creatureTypes.reduce((obj, k) => {
-      obj[k] = CONFIG.DND5E.creatureTypes[k]?.label;
-      return obj;
-    }, {});
     return options;
   }
 
@@ -434,70 +392,5 @@ export default class AbilityUseDialog extends Dialog {
     if ( !hasUses || !uses.autoDestroy ) return false;
     const value = uses.value - consume;
     return value <= 0;
-  }
-
-  /* -------------------------------------------- */
-  /*  Event Handlers                              */
-  /* -------------------------------------------- */
-
-  /** @inheritDoc */
-  activateListeners(jQuery) {
-    super.activateListeners(jQuery);
-    const [html] = jQuery;
-
-    html.querySelector('[name="slotLevel"]')?.addEventListener("change", this._onChangeSlotLevel.bind(this));
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Update summoning profiles when spell slot level is changed.
-   * @param {Event} event  Triggering change event.
-   */
-  _onChangeSlotLevel(event) {
-    const level = this.item.actor?.system.spells?.[event.target.value]?.level;
-    const item = this.item.clone({ "system.level": level ?? this.item.system.level });
-    this._updateProfilesInput(
-      "summonsProfile", "DND5E.Summoning.Profile.Label", this.constructor._createSummoningOptions(item)
-    );
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Update enchantment or summoning profiles inputs when the level changes.
-   * @param {string} name                Name of the field to update.
-   * @param {string} label               Localization key for the field's aria label.
-   * @param {object} options
-   * @param {object} [options.profiles]  Profile options to display, if multiple.
-   * @param {string} [options.profile]   Single profile to select, if only one.
-   */
-  _updateProfilesInput(name, label, options) {
-    const originalInput = this.element[0].querySelector(`[name="${name}"]`);
-    if ( !originalInput ) return;
-
-    // If multiple profiles, replace with select element
-    if ( options.profiles ) {
-      const select = document.createElement("select");
-      select.name = name;
-      select.ariaLabel = game.i18n.localize(label);
-      for ( const [id, label] of Object.entries(options.profiles) ) {
-        const option = document.createElement("option");
-        option.value = id;
-        option.innerText = label;
-        if ( id === originalInput.value ) option.selected = true;
-        select.append(option);
-      }
-      originalInput.replaceWith(select);
-    }
-
-    // If only one profile, replace with hidden input
-    else {
-      const input = document.createElement("input");
-      input.type = "hidden";
-      input.name = name;
-      input.value = options.profile ?? "";
-      originalInput.replaceWith(input);
-    }
   }
 }
