+++ templates/activity/summon-usage-creation.hbs
@@ -1,0 +1,18 @@
+<section>
+    {{> "dnd5e.activity-usage-notes" }}
+    {{#if hasCreation}}
+    <fieldset>
+        <legend>{{ localize "DND5E.USAGE.SECTION.Creation" }}</legend>
+        {{#each summonsFields}}
+        {{ formField field name=name value=value options=options }}
+        {{/each}}
+        {{#if summonsProfile}}
+        <input type="hidden" name="summons.profile" value="{{ summonsProfile }}">
+        {{/if}}
+        {{#if (and summonsFields template)}}<hr>{{/if}}
+        {{#if template}}
+        {{ formField template.field name=template.name value=template.value }}
+        {{/if}}
+    </fieldset>
+    {{/if}}
+</section>
